2.13: Promote ascan rules
