+++ addOns/ascanrules/CHANGELOG.md
@@ -7,6 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - The Format String Error scan rule now includes example alert functionality for documentation generation purposes (Issue 6119).
 - Corrected Hidden File Finder scan rule Blazor WASM config file path.
+- The following scan rules were added, having been promoted from Beta:
+  - Log4Shell
+  - Spring Actuator Information Leak
+  - Spring4Shell
+  - Server Side Template Injection
+  - Server Side Template Injection (Blind)
+  - XPath Injection
 
 ## [55] - 2023-06-06
 ### Changed
