+++ addOns/ascanrules/src/main/java/org/zaproxy/zap/extension/ascanrules/ssti/GoTemplateFormat.java
@@ -17,22 +17,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.zaproxy.zap.extension.ascanrulesBeta.ssti;
+package org.zaproxy.zap.extension.ascanrules.ssti;
 
 /**
- * This represents the code that is necessary to execute an arithmetic operation in Django template
+ * This represents the code that is necessary to execute an arithmetic operation in Golang template
  * engine and the expected result of the operation.
  *
  * @author DiogoMRSilva (2018)
  */
-public class DjangoTemplateFormat extends TemplateFormat {
+public class GoTemplateFormat extends TemplateFormat {
 
-    public DjangoTemplateFormat() {
-        super("{{", "}}", "{{%d0|add:%d0}}");
+    public GoTemplateFormat() {
+        super("{", "}", "{{print \"%d\" \"%d\"}}");
     }
 
     @Override
     public int getExpectedResult(int number1, int number2) {
-        return number1 * 10 + number2 * 10;
+        String concatenated = String.format("%d%d", number1, number2);
+        return Integer.parseInt(concatenated);
     }
 }
