+++ addOns/ascanrules/src/main/java/org/zaproxy/zap/extension/ascanrules/ssti/SinkPoint.java
@@ -17,7 +17,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.zaproxy.zap.extension.ascanrulesBeta.ssti;
+package org.zaproxy.zap.extension.ascanrules.ssti;
 
 import org.parosproxy.paros.network.HttpMessage;
