+++ addOns/ascanrules/src/main/javahelp/org/zaproxy/zap/extension/ascanrules/resources/help/contents/ascanrules.html
@@ -183,6 +183,13 @@ The following is an example JSON entry:
 <p>
 Latest code: <a href="https://github.com/zaproxy/zap-extensions/blob/main/addOns/ascanrules/src/main/java/org/zaproxy/zap/extension/ascanrules/HiddenFilesScanRule.java">HiddenFilesScanRule.java</a>
 
+<H2>Log4Shell (CVE-2021-44228 and CVE-2021-45046)</H2>
+This rule attempts to discover the Log4Shell (<a href="https://nvd.nist.gov/vuln/detail/CVE-2021-44228">CVE-2021-44228</a> and <a href="https://nvd.nist.gov/vuln/detail/CVE-2021-45046">CVE-2021-45046</a>) vulnerabilities.
+It relies on the OAST add-on to generate out-of-band payloads and verify DNS interactions.
+We recommend that this scan rule is used with header injection enabled for maximum coverage.
+<p>
+Latest code: <a href="https://github.com/zaproxy/zap-extensions/blob/main/addOns/ascanrules/src/main/java/org/zaproxy/zap/extension/ascanrules/Log4ShellScanRule.java">Log4ShellScanRule.java</a>
+
 <H2>Padding Oracle</H2>
 This rule attempts to manipulate the padding of encrypted strings to trigger an error response 
 indicating a likely padding oracle vulnerability. 
@@ -232,6 +239,16 @@ a pattern indicating the SSI was successful, an alert is raised and the scanner 
 <p>
 Latest code: <a href="https://github.com/zaproxy/zap-extensions/blob/main/addOns/ascanrules/src/main/java/org/zaproxy/zap/extension/ascanrules/ServerSideIncludeScanRule.java">ServerSideIncludeScanRule.java</a>
 
+<H2>Server Side Template Injection</H2>
+This rule attempts to detect situations in which user input might be interpreted as part of the template and processed on the server, versus the user input simply being used as an argument to the template/engine.
+<p>
+Latest code: <a href="https://github.com/zaproxy/zap-extensions/blob/main/addOns/ascanrules/src/main/java/org/zaproxy/zap/extension/ascanrules/SstiScanRule.java">SstiScanRule.java</a>
+
+<H2>Server Side Template Injection (Blind)</H2>
+This rule goes one step further than the SSTI scan rule and attempts to find places where the impact of the user input is not immediately obvious, such as when used by an admin panel, report output, invoice, etc.
+<p>
+Latest code: <a href="https://github.com/zaproxy/zap-extensions/blob/main/addOns/ascanrules/src/main/java/org/zaproxy/zap/extension/ascanrules/SstiBlindScanRule.java">SstiBlindScanRule.java</a>
+
 <H2>Source Code Disclosure - CVE-2012-1823</H2>
 Exploit CVE-2012-1823 to disclose server-side PHP source code on a PHP-CGI based web server.<br>
 Only analyzes responses that are text based (HTML, JSON, XML, etc.), in order to avoid false positives which may occur with image or other binary content.<br>
@@ -243,6 +260,13 @@ Latest code: <a href="https://github.com/zaproxy/zap-extensions/blob/main/addOns
 Exploit the presence of an unprotected /WEB-INF folder to download and decompile Java classes, to disclose Java source code.
 <p>
 Latest code: <a href="https://github.com/zaproxy/zap-extensions/blob/main/addOns/ascanrules/src/main/java/org/zaproxy/zap/extension/ascanrules/SourceCodeDisclosureWebInfScanRule.java">SourceCodeDisclosureWebInfScanRule.java</a>
+
+<H2>Spring4Shell (CVE-2022-22965)</H2>
+This rule attempts to discover the Spring4Shell (<a href="https://tanzu.vmware.com/security/cve-2022-22965">CVE-2022-22965</a> vulnerability.
+It uses a payload of <code>class.module.classLoader.DefaultAssertionStatus=nonsense</code> on all nodes and raises an alert if this payload results in a 400 response.
+It will not raise an alert if a similar but safe payload also results in a 400 response.
+<p>
+Latest code: <a href="https://github.com/zaproxy/zap-extensions/blob/main/addOns/ascanrules/src/main/java/org/zaproxy/zap/extension/ascanrules/Spring4ShellScanRule.java">Spring4ShellScanRule.java</a>
 
 <H2>SQL Injection</H2>
 
@@ -341,6 +365,16 @@ This active scan rule checks for differences in response based on fuzzed User Ag
 <strong>Note:</strong> If the Custom Payloads addon is installed you can add your own User Agent strings (payloads) in the Custom Payloads options panel.
 <p>
 Latest code: <a href="https://github.com/zaproxy/zap-extensions/blob/main/addOns/ascanrules/src/main/java/org/zaproxy/zap/extension/ascanrules/UserAgentScanRule.java">UserAgentScanRule.java</a>
+
+<H2>XPath Injection</H2>
+As described by OWASP: "XPath Injection attacks occur when a web site uses user-supplied information to 
+construct an XPath query for XML data. By sending intentionally malformed information into the web site, 
+an attacker can find out how the XML data is structured, or access data that he may not normally have access 
+to. He may even be able to elevate his privileges on the web site if the XML data is being used for 
+authentication (such as an XML based user file) or authorization." This rule attempts to identify such 
+weaknesses.
+<p>
+Latest code: <a href="https://github.com/zaproxy/zap-extensions/blob/main/addOns/ascanrules/src/main/java/org/zaproxy/zap/extension/ascanrules/XpathInjectionScanRule.java">XpathInjectionScanRule.java</a>
 
 <H2>XSLT Injection</H2>
 This scan rule checks for certain responses induced by injecting XSL transformations. <br/>
