+++ addOns/ascanrules/src/main/resources/org/zaproxy/zap/extension/ascanrules/resources/Messages.properties
@@ -93,6 +93,17 @@ ascanrules.htaccess.otherinfo = Based on response status code htaccess file may 
 ascanrules.htaccess.refs = http://www.htaccess-guide.com/
 ascanrules.htaccess.soln = Ensure the .htaccess file is not accessible.
 
+ascanrules.log4shell.cve44228.desc = Apache Log4j2 <=2.14.1 JNDI features used in configuration, log messages, and parameters do not protect against attacker controlled LDAP and other JNDI related endpoints. An attacker who can control log messages or log message parameters can execute arbitrary code loaded from LDAP servers when message lookup substitution is enabled. From log4j 2.15.0, this behavior has been disabled by default.
+ascanrules.log4shell.cve44228.name = Log4Shell (CVE-2021-44228)
+ascanrules.log4shell.cve44228.refs = https://www.lunasec.io/docs/blog/log4j-zero-day/\nhttps://nvd.nist.gov/vuln/detail/CVE-2021-44228
+ascanrules.log4shell.cve44228.soln = Upgrade Log4j2 to version 2.17.1 or newer. In previous releases (>2.10) this behavior can be mitigated by setting system property "log4j2.formatMsgNoLookups" to "true" or by removing the JndiLookup class from the classpath (example: zip -q -d log4j-core-*.jar org/apache/logging/log4j/core/lookup/JndiLookup.class). Java 8u121 (see https://www.oracle.com/java/technologies/javase/8u121-relnotes.html) protects against remote code execution by defaulting "com.sun.jndi.rmi.object.trustURLCodebase" and "com.sun.jndi.cosnaming.object.trustURLCodebase" to "false".
+ascanrules.log4shell.cve45046.desc = It was found that the fix to address CVE-2021-44228 in Apache Log4j 2.15.0 was incomplete in certain non-default configurations. This could allow attackers to craft malicious input data using a JNDI Lookup pattern resulting in an information leak and remote code execution in some environments.
+ascanrules.log4shell.cve45046.name = Log4Shell (CVE-2021-45046)
+ascanrules.log4shell.cve45046.refs = https://www.lunasec.io/docs/blog/log4j-zero-day/\nhttps://nvd.nist.gov/vuln/detail/CVE-2021-45046
+ascanrules.log4shell.cve45046.soln = Upgrade Log4j2 to version 2.17.1 or newer.
+ascanrules.log4shell.name = Log4Shell
+ascanrules.log4shell.skipped = no Active Scan OAST service is selected.
+
 ascanrules.name = Active Scan Rules
 
 ascanrules.paddingoracle.desc = By manipulating the padding on an encrypted string, an attacker is able to generate an error message that indicates a likely 'padding oracle' vulnerability. Such a vulnerability can affect any application or framework that uses encryption improperly, such as some versions of ASP.net, Java Server Faces, and Mono. An attacker may exploit this issue to decrypt data and recover encryption keys, potentially viewing and modifying confidential data. This rule should detect the MS10-070 padding oracle vulnerability in ASP.net if CustomErrors are enabled for that.
@@ -144,6 +155,16 @@ ascanrules.sourcecodedisclosurewebinf.propertiesfile.name = Properties File Disc
 ascanrules.sourcecodedisclosurewebinf.propertiesfile.soln = The web server should be configured to not serve the /WEB-INF folder or its contents to web browsers.  It may also be possible to remove the /WEB-INF folder.  
 ascanrules.sourcecodedisclosurewebinf.soln = The web server should be configured to not serve the /WEB-INF folder or its contents to web browsers, since it contains sensitive information such as compiled Java source code and properties files which may contain credentials. Java classes deployed with the application should be obfuscated, as an additional layer of defence in a "defence-in-depth" approach.
 
+ascanrules.spring4shell.desc = The application appears to be vulnerable to CVE-2022-22965 (otherwise known as Spring4Shell) - remote code execution (RCE) via data binding.
+ascanrules.spring4shell.name = Spring4Shell
+ascanrules.spring4shell.refs = https://nvd.nist.gov/vuln/detail/CVE-2022-22965\nhttps://www.rapid7.com/blog/post/2022/03/30/spring4shell-zero-day-vulnerability-in-spring-framework/\nhttps://spring.io/blog/2022/03/31/spring-framework-rce-early-announcement#vulnerability\nhttps://tanzu.vmware.com/security/cve-2022-22965
+ascanrules.spring4shell.soln = Upgrade Spring Framework to versions 5.3.18, 5.2.20, or newer.
+
+ascanrules.springactuator.desc = Spring Actuator for Health is enabled and may reveal sensitive information about this application. Spring Actuators can be used for real monitoring purposes, but should be used with caution as to not expose too much information about the application or the infrastructure running it.
+ascanrules.springactuator.name = Spring Actuator Information Leak
+ascanrules.springactuator.refs = https://docs.spring.io/spring-boot/docs/current/actuator-api/htmlsingle/#overview
+ascanrules.springactuator.soln = Disable the Health Actuators and other actuators, or restrict them to administrative users.
+
 #ascanrules.sqlinjection.alert.errorbased.attack={1}
 ascanrules.sqlinjection.alert.booleanbased.attack = field: [{0}], value [{1}]
 ascanrules.sqlinjection.alert.booleanbased.extrainfo = The page results were successfully manipulated using the boolean conditions [{0}] and [{1}]\nThe parameter value being modified was {2}stripped from the HTML output for the purposes of the comparison
@@ -176,6 +197,18 @@ ascanrules.sqlinjection.sqlite.alert.timebased.extrainfo = The query time is con
 ascanrules.sqlinjection.sqlite.alert.versionnumber.extrainfo = Using a UNION based SQL Injection attack, and by exploiting SQLite's dynamic typing mechanism, the SQLite version was determined to be [{0}].\nWith string-based injection points, full SQLite version information can be extracted, but with numeric injection points, only partial SQLite version information can be extracted.\nMore information on SQLite version [{0}] is available at http://www.sqlite.org/changes.html
 ascanrules.sqlinjection.sqlite.name = SQL Injection - SQLite
 
+ascanrules.ssti.alert.otherinfo = Proof found at [{0}] \ncontent:\n[{1}]
+ascanrules.ssti.desc = When the user input is inserted in the template instead of being used as argument in rendering is evaluated by the template engine. Depending on the template engine it can lead to remote code execution.
+ascanrules.ssti.name = Server Side Template Injection
+ascanrules.ssti.refs = https://portswigger.net/blog/server-side-template-injection
+ascanrules.ssti.soln = Instead of inserting the user input in the template, use it as rendering argument.
+
+ascanrules.sstiblind.alert.recvdcallback.otherinfo = Received callback from the server.
+ascanrules.sstiblind.desc = When the user input is inserted in the template instead of being used as argument in rendering is evaluated by the template engine. Depending on the template engine it can lead to remote code execution.
+ascanrules.sstiblind.name = Server Side Template Injection (Blind)
+ascanrules.sstiblind.refs = https://portswigger.net/blog/server-side-template-injection
+ascanrules.sstiblind.soln = Instead of inserting the user input in the template, use it as rendering argument.
+
 ascanrules.traceaxd.desc = The ASP.NET Trace Viewer (trace.axd) was found to be available. This component can leak a significant amount of valuable information.
 ascanrules.traceaxd.name = Trace.axd Information Leak
 ascanrules.traceaxd.otherinfo = Based on response status code Trace Viewer may be protected by an authentication or authorization mechanism.  
@@ -186,6 +219,8 @@ ascanrules.useragent.desc = Check for differences in response based on fuzzed Us
 ascanrules.useragent.name = User Agent Fuzzer
 ascanrules.useragent.refs = https://owasp.org/wstg
 ascanrules.useragent.useragentparmname = Header User-Agent
+
+ascanrules.xpathinjection.name = XPath Injection
 
 ascanrules.xsltinjection.command.otherinfo = Command execution may be possible.
 ascanrules.xsltinjection.desc = Injection using XSL transformations may be possible, and may allow an attacker to read system information, read and write files, or execute arbitrary code.
