+++ addOns/ascanrules/src/test/resources/org/zaproxy/zap/extension/ascanrules/sstiscanrule/Rendered.html
