+++ addOns/ascanrulesAlpha/src/main/resources/org/zaproxy/zap/extension/ascanrulesAlpha/resources/Messages.properties
@@ -31,19 +31,6 @@ ascanalpha.mongodb.soln = Do not trust client side input and escape all data on 
 
 ascanalpha.name = Active Scan Rules - alpha
 
-ascanalpha.ssrf.desc = The web server receives a remote address and retrieves the contents of this URL, but it does not sufficiently ensure that the request is being sent to the expected destination.
-ascanalpha.ssrf.name = Server Side Request Forgery
-ascanalpha.ssrf.otherinfo.canaryinbody = The canary token from the out-of-band service was found in the response body.
-ascanalpha.ssrf.refs = https://cheatsheetseries.owasp.org/cheatsheets/Server_Side_Request_Forgery_Prevention_Cheat_Sheet.html
-ascanalpha.ssrf.skipped = no Active Scan OAST service is selected.
-ascanalpha.ssrf.soln = Do not accept remote addresses as request parameters, and if you must, ensure that they are validated against an allow-list of expected values.
-
-ascanalpha.text4shell.desc = Apache Commons Text prior to 1.10.0 allows RCE when applied to untrusted input due to insecure interpolation defaults.Apache Commons Text performs variable interpolation, allowing properties to be dynamically evaluated and expanded.The application has been shown to initial contact with remote servers via variable interpolation and may well be vulnerable to Remote Code Execution (RCE).
-ascanalpha.text4shell.name = Text4shell (CVE-2022-42889)
-ascanalpha.text4shell.refs = https://nvd.nist.gov/vuln/detail/CVE-2022-42889\nhttps://securitylab.github.com/advisories/GHSL-2022-018_Apache_Commons_Text/
-ascanalpha.text4shell.skipped = no Active Scan OAST service is selected.
-ascanalpha.text4shell.soln = Upgrade Apache Commons Text prior to version 1.10.0 or newer.
-
 ascanalpha.webCacheDeception.desc = Web cache deception may be possible. It may be possible for unauthorised user to view sensitive data on this page.
 ascanalpha.webCacheDeception.name = Web Cache Deception
 ascanalpha.webCacheDeception.otherinfo = Cached Authorised Response and Unauthorised Response are similar.
